backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
joblib
scikit-learn
xgboost
supabase
python-dotenv
pandas
randomforest